初步主体结构梳理

+++ script/plugin/oipage-time-plugin.js
@@ -10,7 +10,6 @@ class OIPageTimePlugin {
         compiler.hooks.beforeRun.tap(
             pluginName, () => {
 
-                // 如果缓存文件夹不存在，创建
                 if (fs.existsSync(process.env.OIPAGE_CACHE)) {
                     deleteSync(process.env.OIPAGE_CACHE);
                 }
@@ -22,12 +21,14 @@ class OIPageTimePlugin {
             pluginName, () => {
 
                 if (process.env.OIPAGE_PLATFORM != "h5") {
-                    listFileSync(process.env.OIPAGE_CACHE, function (filepath) {
-                        copySync(filepath.path, filepath.path.replace(process.env.OIPAGE_CACHE, process.env.OIPAGE_DIST));
-                    });
-                    deleteSync(process.env.OIPAGE_CACHE);
+                    if (fs.existsSync(process.env.OIPAGE_CACHE)) {
+                        listFileSync(process.env.OIPAGE_CACHE, function (filepath) {
+                            copySync(filepath.path, filepath.path.replace(process.env.OIPAGE_CACHE, process.env.OIPAGE_DIST));
+                        });
+                        deleteSync(process.env.OIPAGE_CACHE);
+                    }
                 }
-                
+
             }
         );
